% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{antichain}
\alias{antichain}
\title{Antichain}
\usage{
antichain(n)
}
\arguments{
\item{n}{Number of incomparable elements}
}
\value{
An antichain.
}
\description{
Antichain
}
\examples{
A <- antichain(3)
}
