% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryRelation.R
\name{BinaryRelation}
\alias{BinaryRelation}
\title{Binary Relation}
\description{
Binary Relation

Binary Relation
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BinaryRelation-new}{\code{BinaryRelation$new()}}
\item \href{#method-BinaryRelation-print}{\code{BinaryRelation$print()}}
\item \href{#method-BinaryRelation-clone}{\code{BinaryRelation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryRelation-new"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryRelation-new}{}}}
\subsection{Method \code{new()}}{
Constructor of a Binary Relation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryRelation$new(matrix = NULL, elements = NULL, file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{A squared matrix indicating the binary relation between elements.}

\item{\code{elements}}{A vector with element names.}

\item{\code{file}}{If provided, a file to load.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The BinaryRelation object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryRelation-print"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryRelation-print}{}}}
\subsection{Method \code{print()}}{
Prints the BinaryRelation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryRelation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinaryRelation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BinaryRelation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryRelation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
