% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{boole}
\alias{boole}
\title{Boole Algebra}
\usage{
boole(n)
}
\arguments{
\item{n}{Number of atoms in the Boole algebra}
}
\value{
A Lattice object representing the Boole algebra with n atoms.
}
\description{
Boole Algebra
}
\examples{
B <- boole(2)
}
