% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{chain}
\alias{chain}
\title{Chain}
\usage{
chain(...)
}
\arguments{
\item{...}{The number of elements (single input) or the list of elements}
}
\value{
A chain lattice with the given configuration.
}
\description{
Chain
}
\examples{
C1 <- chain(7)
C2 <- chain("a", "b", "c")
}
