% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{divisors_lattice}
\alias{divisors_lattice}
\title{Lattice of divisors}
\usage{
divisors_lattice(n)
}
\arguments{
\item{n}{The number in the top of the lattice}
}
\value{
A Lattice object
}
\description{
Lattice of divisors
}
\examples{
P <- divisors_lattice(60)
}
