% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_conceptlattice.R
\name{from_conceptlattice}
\alias{from_conceptlattice}
\title{Create a lattice from a concept lattice.}
\usage{
from_conceptlattice(concepts)
}
\arguments{
\item{concepts}{A concept lattice object.}
}
\value{
A lattice object.
}
\description{
Create a lattice from a concept lattice.
}
