% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product.R
\name{\%p\%}
\alias{\%p\%}
\title{Product of two posets}
\usage{
P1 \%p\% P2
}
\arguments{
\item{P1}{The first poset.}

\item{P2}{The second poset.}
}
\value{
The product poset.
}
\description{
Computes the product of two posets.
}
\examples{
P1 <- Poset$new(matrix(c(0, 1, 0, 1), ncol = 2))
P2 <- Poset$new(matrix(c(0, 1, 0, 0), ncol = 2))
P <- P1 \%p\% P2

}
