% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lattice.R
\name{Lattice}
\alias{Lattice}
\title{R6 class for a lattice}
\description{
This class implements the data structure and methods for lattices.
}
\section{Super class}{
\code{\link[LatticeTheory:Poset]{LatticeTheory::Poset}} -> \code{Lattice}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lattice-print}{\code{Lattice$print()}}
\item \href{#method-Lattice-top}{\code{Lattice$top()}}
\item \href{#method-Lattice-bottom}{\code{Lattice$bottom()}}
\item \href{#method-Lattice-join_irreducibles}{\code{Lattice$join_irreducibles()}}
\item \href{#method-Lattice-meet_irreducibles}{\code{Lattice$meet_irreducibles()}}
\item \href{#method-Lattice-atoms}{\code{Lattice$atoms()}}
\item \href{#method-Lattice-coatoms}{\code{Lattice$coatoms()}}
\item \href{#method-Lattice-complements}{\code{Lattice$complements()}}
\item \href{#method-Lattice-interval}{\code{Lattice$interval()}}
\item \href{#method-Lattice-chains}{\code{Lattice$chains()}}
\item \href{#method-Lattice-context}{\code{Lattice$context()}}
\item \href{#method-Lattice-dual}{\code{Lattice$dual()}}
\item \href{#method-Lattice-clone}{\code{Lattice$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="as_graph"><a href='../../LatticeTheory/html/Poset.html#method-Poset-as_graph'><code>LatticeTheory::Poset$as_graph()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="infimum"><a href='../../LatticeTheory/html/Poset.html#method-Poset-infimum'><code>LatticeTheory::Poset$infimum()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="initialize"><a href='../../LatticeTheory/html/Poset.html#method-Poset-initialize'><code>LatticeTheory::Poset$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="is_lower_than"><a href='../../LatticeTheory/html/Poset.html#method-Poset-is_lower_than'><code>LatticeTheory::Poset$is_lower_than()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="latex"><a href='../../LatticeTheory/html/Poset.html#method-Poset-latex'><code>LatticeTheory::Poset$latex()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="linear_extension"><a href='../../LatticeTheory/html/Poset.html#method-Poset-linear_extension'><code>LatticeTheory::Poset$linear_extension()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="lower_bounds"><a href='../../LatticeTheory/html/Poset.html#method-Poset-lower_bounds'><code>LatticeTheory::Poset$lower_bounds()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="lower_cone"><a href='../../LatticeTheory/html/Poset.html#method-Poset-lower_cone'><code>LatticeTheory::Poset$lower_cone()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="lower_neighbours"><a href='../../LatticeTheory/html/Poset.html#method-Poset-lower_neighbours'><code>LatticeTheory::Poset$lower_neighbours()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="maximals"><a href='../../LatticeTheory/html/Poset.html#method-Poset-maximals'><code>LatticeTheory::Poset$maximals()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="maximum"><a href='../../LatticeTheory/html/Poset.html#method-Poset-maximum'><code>LatticeTheory::Poset$maximum()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="minimals"><a href='../../LatticeTheory/html/Poset.html#method-Poset-minimals'><code>LatticeTheory::Poset$minimals()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="minimum"><a href='../../LatticeTheory/html/Poset.html#method-Poset-minimum'><code>LatticeTheory::Poset$minimum()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="plot"><a href='../../LatticeTheory/html/Poset.html#method-Poset-plot'><code>LatticeTheory::Poset$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="size"><a href='../../LatticeTheory/html/Poset.html#method-Poset-size'><code>LatticeTheory::Poset$size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="supremum"><a href='../../LatticeTheory/html/Poset.html#method-Poset-supremum'><code>LatticeTheory::Poset$supremum()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="upper_bounds"><a href='../../LatticeTheory/html/Poset.html#method-Poset-upper_bounds'><code>LatticeTheory::Poset$upper_bounds()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="upper_cone"><a href='../../LatticeTheory/html/Poset.html#method-Poset-upper_cone'><code>LatticeTheory::Poset$upper_cone()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LatticeTheory" data-topic="Poset" data-id="upper_neighbours"><a href='../../LatticeTheory/html/Poset.html#method-Poset-upper_neighbours'><code>LatticeTheory::Poset$upper_neighbours()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-print"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-print}{}}}
\subsection{Method \code{print()}}{
Print a Lattice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-top"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-top}{}}}
\subsection{Method \code{top()}}{
Top of a Lattice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$top()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The top of the Concept Lattice
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-bottom"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-bottom}{}}}
\subsection{Method \code{bottom()}}{
Bottom of a Lattice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$bottom()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The bottom of the Concept Lattice
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-join_irreducibles"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-join_irreducibles}{}}}
\subsection{Method \code{join_irreducibles()}}{
Join-irreducible Elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$join_irreducibles()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The join-irreducible elements in the lattice.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-meet_irreducibles"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-meet_irreducibles}{}}}
\subsection{Method \code{meet_irreducibles()}}{
Meet-irreducible Elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$meet_irreducibles()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The meet-irreducible elements in the lattice.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-atoms"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-atoms}{}}}
\subsection{Method \code{atoms()}}{
Atoms of a lattice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$atoms()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The atoms of the lattice.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-coatoms"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-coatoms}{}}}
\subsection{Method \code{coatoms()}}{
Coatoms of a lattice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$coatoms()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The coatoms (superatoms) of the lattice.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-complements"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-complements}{}}}
\subsection{Method \code{complements()}}{
Complement(s) of a lattice element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$complements(a)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{Name of an element}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The name of the complement elements, if any; \code{NULL}, otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-interval"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-interval}{}}}
\subsection{Method \code{interval()}}{
Interval in a lattice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$interval(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{First (lower) element}

\item{\code{b}}{Secod (greater) element}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The elements which are greater than or equal to \code{a} and lower than or equal to \code{b}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-chains"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-chains}{}}}
\subsection{Method \code{chains()}}{
Chains in a lattice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$chains(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{First (lower) element}

\item{\code{b}}{Secod (greater) element}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The chains starting at \code{a} and ending in \code{b}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-context"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-context}{}}}
\subsection{Method \code{context()}}{
Standard context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$context()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A formal context (in \code{fcaR} is installed) or a matrix of a context whose associated concept lattice is (isomorphic to) the current lattice.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-dual"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-dual}{}}}
\subsection{Method \code{dual()}}{
Dual of the lattice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$dual()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Another lattice where the order is reversed.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lattice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lattice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lattice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
