% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatticeTheory-package.R
\docType{package}
\name{LatticeTheory-package}
\alias{LatticeTheory}
\alias{LatticeTheory-package}
\title{LatticeTheory: Managing of lattices}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
