% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poset.R
\name{Poset}
\alias{Poset}
\title{R6 class for a poset}
\description{
This class implements the data structure and methods for partial ordered sets.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{names}}{Names of the elements.}

\item{\code{order}}{Matrix representing the dominance.}

\item{\code{reduced_matrix}}{Matrix representing dominance, once transitivity is taken into account.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Poset-new}{\code{Poset$new()}}
\item \href{#method-Poset-print}{\code{Poset$print()}}
\item \href{#method-Poset-size}{\code{Poset$size()}}
\item \href{#method-Poset-maximals}{\code{Poset$maximals()}}
\item \href{#method-Poset-maximum}{\code{Poset$maximum()}}
\item \href{#method-Poset-minimals}{\code{Poset$minimals()}}
\item \href{#method-Poset-minimum}{\code{Poset$minimum()}}
\item \href{#method-Poset-supremum}{\code{Poset$supremum()}}
\item \href{#method-Poset-infimum}{\code{Poset$infimum()}}
\item \href{#method-Poset-upper_bounds}{\code{Poset$upper_bounds()}}
\item \href{#method-Poset-lower_bounds}{\code{Poset$lower_bounds()}}
\item \href{#method-Poset-lower_cone}{\code{Poset$lower_cone()}}
\item \href{#method-Poset-upper_cone}{\code{Poset$upper_cone()}}
\item \href{#method-Poset-lower_neighbours}{\code{Poset$lower_neighbours()}}
\item \href{#method-Poset-upper_neighbours}{\code{Poset$upper_neighbours()}}
\item \href{#method-Poset-is_lower_than}{\code{Poset$is_lower_than()}}
\item \href{#method-Poset-linear_extension}{\code{Poset$linear_extension()}}
\item \href{#method-Poset-to_lattice}{\code{Poset$to_lattice()}}
\item \href{#method-Poset-plot}{\code{Poset$plot()}}
\item \href{#method-Poset-latex}{\code{Poset$latex()}}
\item \href{#method-Poset-as_graph}{\code{Poset$as_graph()}}
\item \href{#method-Poset-to_context}{\code{Poset$to_context()}}
\item \href{#method-Poset-completion}{\code{Poset$completion()}}
\item \href{#method-Poset-clone}{\code{Poset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-new"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-new}{}}}
\subsection{Method \code{new()}}{
Initialize a poset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$new(order, names = colnames(order))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{order}}{Matrix indicating the order relationship}

\item{\code{names}}{Character vector with names of the elements}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class Poset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-print"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-print}{}}}
\subsection{Method \code{print()}}{
Print a Poset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-size"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-size}{}}}
\subsection{Method \code{size()}}{
Size of poset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of elements in the poset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-maximals"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-maximals}{}}}
\subsection{Method \code{maximals()}}{
Maximals of elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$maximals(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{element names (as vector or enumerated).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The maximals of the specified elements
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-maximum"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-maximum}{}}}
\subsection{Method \code{maximum()}}{
Maximum of elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$maximum(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{element names (as vector or enumerated).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The maximum of the specified elements, if it exists; \code{NULL} if it does not exist.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-minimals"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-minimals}{}}}
\subsection{Method \code{minimals()}}{
Minimals of elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$minimals(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{element names (as vector or enumerated).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The minimals of the specified elements.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-minimum"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-minimum}{}}}
\subsection{Method \code{minimum()}}{
Minimum of elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$minimum(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{element names (as vector or enumerated).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The minimum of the specified elements, if it exists; \code{NULL} if it does not exist.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-supremum"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-supremum}{}}}
\subsection{Method \code{supremum()}}{
Supremum of elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$supremum(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Element names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The supremum of the list of elements (if it exists); \code{NULL}, otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-infimum"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-infimum}{}}}
\subsection{Method \code{infimum()}}{
Infimum of elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$infimum(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Element names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The infimum of the list of elements (if it exists); \code{NULL}, otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-upper_bounds"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-upper_bounds}{}}}
\subsection{Method \code{upper_bounds()}}{
Upper bound of elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$upper_bounds(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Element names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The set of upper bounds of the list of elements.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-lower_bounds"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-lower_bounds}{}}}
\subsection{Method \code{lower_bounds()}}{
Lower bound of elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$lower_bounds(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Element names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The set of lower bounds of the list of elements.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-lower_cone"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-lower_cone}{}}}
\subsection{Method \code{lower_cone()}}{
Lower cone of an element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$lower_cone(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The name of an element.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector with the elements of the lower cone (that is, its lower bounds).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-upper_cone"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-upper_cone}{}}}
\subsection{Method \code{upper_cone()}}{
Upper cone of an element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$upper_cone(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The name of an element.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector with the elements of the upper cone (that is, its upper bounds).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-lower_neighbours"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-lower_neighbours}{}}}
\subsection{Method \code{lower_neighbours()}}{
Lower Neighbours of an element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$lower_neighbours(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The name of an element}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the lower neighbours of element \code{id}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-upper_neighbours"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-upper_neighbours}{}}}
\subsection{Method \code{upper_neighbours()}}{
Upper Neighbours of an element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$upper_neighbours(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The name of an element}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the upper neighbours of element \code{id}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-is_lower_than"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-is_lower_than}{}}}
\subsection{Method \code{is_lower_than()}}{
Order of two elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$is_lower_than(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{The name of an element}

\item{\code{b}}{The name of the other element}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if \code{a} is lower than or equal to \code{b}; \code{FALSE}, otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-linear_extension"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-linear_extension}{}}}
\subsection{Method \code{linear_extension()}}{
Linear extension of poset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$linear_extension()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Another poset which is a linear extension of this.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-to_lattice"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-to_lattice}{}}}
\subsection{Method \code{to_lattice()}}{
Convert to Lattice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$to_lattice()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A Lattice object if this poset is actually a lattice, otherwise it is returned unmodified.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-plot}{}}}
\subsection{Method \code{plot()}}{
Hasse diagram of Poset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$plot(engine = c("own", "hasseDiagram", "POSetR"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{engine}}{Which engine to use for the plot?}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Available engines are \code{own} (uses ggplot2), \code{hasseDiagram} (uses the hasseDiagram package, which should be installed); and \code{POSetR}, which uses the POSetR package.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-latex"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-latex}{}}}
\subsection{Method \code{latex()}}{
Hasse diagram of Poset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$latex(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Other arguments.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The LaTeX code to reproduce the Hasse diagram of the poset.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-as_graph"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-as_graph}{}}}
\subsection{Method \code{as_graph()}}{
Poset as graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$as_graph()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An \code{igraph} object representing this poset.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-to_context"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-to_context}{}}}
\subsection{Method \code{to_context()}}{
Context related to Poset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$to_context()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Formal Context object representing the incidence relation $(P, P, \le)$ where $P$ are the elements of this Poset. If the fcaR library is present, it returns a \code{FormalContext} object, otherwise, it simply returns the relation matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-completion"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-completion}{}}}
\subsection{Method \code{completion()}}{
Completion of the Poset to Lattice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$completion()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Dedekind-MacNeill completion of the Poset, as a \code{Lattice} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Poset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
