% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatticeTheory-package.R
\docType{package}
\name{LatticeTheory-package}
\alias{LatticeTheory}
\alias{LatticeTheory-package}
\title{LatticeTheory: Managing of lattices}
\description{
Functions and data structures to work with posets and lattices.
}
\author{
\strong{Maintainer}: Domingo Lopez Rodriguez \email{dominlopez@uma.es} (\href{https://orcid.org/0000-0002-0172-1585}{ORCID})

Authors:
\itemize{
  \item Manuel Ojeda Hernández \email{manuojeda@uma.es}
}

}
\keyword{internal}
